# server/bridge/docs/openapi.yaml
openapi: 3.0.0
info:
  title: ELYVN Bridge API
  description: AI-powered lead qualification, call management, and revenue analytics API.
  version: 1.0.0
  contact:
    name: ELYVN Support
servers:
  - url: https://api.elyvn.ai
    description: Production
  - url: http://localhost:3001
    description: Development (default port)
  - url: http://localhost:3000
    description: Development (alternate port)
security:
  - apiKey: []
  - bearerAuth: []
paths:
...
  /bookings/{clientId}:
    get:
      tags: [Bookings]
      summary: List Cal.com bookings for a client
      parameters:
        - $ref: '#/components/parameters/clientId'
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Booking list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object

  # --- Stats ---
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    clientId:
      name: clientId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...